+++ R/bayesfactor.R
@@ -46,15 +46,15 @@
 #' @export
 bayesfactor <-
   function(...,
-             prior = NULL,
-             direction = "two-sided",
-             null = 0,
-             hypothesis = NULL,
-             effects = c("fixed", "random", "all"),
-             verbose = TRUE,
-             denominator = 1,
-             match_models = FALSE,
-             prior_odds = NULL) {
+           prior = NULL,
+           direction = "two-sided",
+           null = 0,
+           hypothesis = NULL,
+           effects = c("fixed", "random", "all"),
+           verbose = TRUE,
+           denominator = 1,
+           match_models = FALSE,
+           prior_odds = NULL) {
     mods <- list(...)
 
     if (length(mods) > 1) {