+++ R/bayesfactor_restricted.R
@@ -3,6 +3,8 @@
 #' This method computes Bayes factors for comparing a model with an order restrictions on its parameters
 #' with the fully unrestricted model. \emph{Note that this method should only be used for confirmatory analyses}.
 #' \cr \cr
+#' The \code{bf_*} function is an alias of the main function.
+#' \cr \cr
 #' \strong{For info on specifying correct priors for factors with more than 2 levels, see \href{https://easystats.github.io/bayestestR/articles/bayes_factors.html}{the Bayes factors vignette}.}
 #' \cr \cr
 #' For more info, see \href{https://easystats.github.io/bayestestR/articles/bayes_factors.html}{the Bayes factors vignette}.
@@ -118,6 +120,10 @@ bayesfactor_restricted <- function(posterior, hypothesis, prior = NULL, verbose 
   UseMethod("bayesfactor_restricted")
 }
 
+#' @rdname bayesfactor_restricted
+#' @export
+bf_restricted <- bayesfactor_restricted
+
 #' @importFrom insight get_parameters
 #' @rdname bayesfactor_restricted
 #' @export
@@ -148,7 +154,6 @@ bayesfactor_restricted.stanreg <- function(posterior, hypothesis, prior = NULL,
 bayesfactor_restricted.brmsfit <- bayesfactor_restricted.stanreg
 
 #' @importFrom stats update
-#' @importFrom insight get_parameters
 #' @rdname bayesfactor_restricted
 #' @export
 bayesfactor_restricted.emmGrid <- function(posterior, hypothesis, prior = NULL,
@@ -166,8 +171,9 @@ bayesfactor_restricted.emmGrid <- function(posterior, hypothesis, prior = NULL,
     )
   } else if (!inherits(prior, "emmGrid")) { # then is it a model
     prior <- .update_to_priors(prior, verbose = verbose)
-    prior <- insight::get_parameters(prior, effects = "fixed")
-    prior <- stats::update(posterior, post.beta = as.matrix(prior))
+    prior <- emmeans::ref_grid(prior)
+    prior <- prior@post.beta
+    prior <- stats::update(posterior, post.beta = prior)
   }
 
   prior <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(prior, names = FALSE)))