+++ R/eti.R
@@ -78,6 +78,16 @@ eti.MCMCglmm <- function(x, ci = .89, verbose = TRUE, ...) {
 
 
 
+#' @export
+eti.mcmc <- function(x, ci = .89, verbose = TRUE, ...) {
+  d <- as.data.frame(x)
+  dat <- .compute_interval_dataframe(x = d, ci = ci, verbose = verbose, fun = "eti")
+  attr(dat, "data") <- deparse(substitute(x), width.cutoff = 500)
+  dat
+}
+
+
+
 #' @rdname eti
 #' @export
 eti.sim.merMod <- function(x, ci = .89, effects = c("fixed", "random", "all"), parameters = NULL, verbose = TRUE, ...) {