+++ R/rope_range.R
@@ -74,47 +74,44 @@ rope_range.lm <- rope_range.brmsfit
 rope_range.merMod <- rope_range.brmsfit
 
 #' @export
-rope_range.sim.merMod <- function(x, ...) {
+rope_range.default <- function(x, ...) {
   c(-.1, .1)
 }
 
-#' @export
-rope_range.sim <- rope_range.sim.merMod
-
-
 .rope_range <- function(x, information, response) {
-  negligible_value <- tryCatch({
-    # Linear Models
-    if (information$is_linear) {
-      0.1 * stats::sd(response)
+  negligible_value <- tryCatch(
+    {
+      # Linear Models
+      if (information$is_linear) {
+        0.1 * stats::sd(response)
 
-      # General Linear Models
-    } else if (information$is_binomial) {
-      0.1 * pi / sqrt(3)
+        # General Linear Models
+      } else if (information$is_binomial) {
+        0.1 * pi / sqrt(3)
 
-      # T-tests
-    } else if (information$is_ttest) {
-      if ("BFBayesFactor" %in% class(x)) {
-        0.1 * stats::sd(x@data[, 1])
+        # T-tests
+      } else if (information$is_ttest) {
+        if ("BFBayesFactor" %in% class(x)) {
+          0.1 * stats::sd(x@data[, 1])
+        } else {
+          warning("Could not estimate a good default ROPE range. Using 'c(-0.1, 0.1)'.", call. = FALSE)
+          0.1
+        }
+
+        # Correlations
+      } else if (information$is_correlation) {
+        # https://github.com/easystats/bayestestR/issues/121
+        0.05
+
+        # Default
       } else {
-        warning("Could not estimate a good default ROPE range. Using 'c(-0.1, 0.1)'.", call. = FALSE)
         0.1
       }
-
-      # Correlations
-    } else if (information$is_correlation) {
-      # https://github.com/easystats/bayestestR/issues/121
-      0.05
-
-      # Default
-    } else {
+    },
+    error = function(e) {
+      warning("Could not estimate a good default ROPE range. Using 'c(-0.1, 0.1)'.", call. = FALSE)
       0.1
     }
-  },
-  error = function(e) {
-    warning("Could not estimate a good default ROPE range. Using 'c(-0.1, 0.1)'.", call. = FALSE)
-    0.1
-  }
   )
 
   c(-1, 1) * negligible_value