+++ tests/testthat/test-emmGrid.R
@@ -10,6 +10,8 @@ c_ <- pairs(em_)
 all_ <- rbind(em_, c_)
 all_summ <- summary(all_)
 
+
+
 test_that("emmGrid ci", {
   testthat::skip_on_travis()
   xci <- ci(all_, ci = 0.9)
@@ -52,10 +54,10 @@ test_that("emmGrid p_map", {
   testthat::expect_equal(xpmap$p_MAP, c(0.42, 0, 0.26), tolerance = 0.1)
 })
 
-test_that("emmGrid p_rope", {
+test_that("emmGrid mhdior", {
   testthat::skip_on_travis()
-  xprope <- p_rope(all_, range = c(-0.1, 0.1), precision = 0.5)
-  testthat::expect_equal(xprope$p_ROPE, c(0.695, 1, 0.87), tolerance = 0.1)
+  xprope <- mhdior(all_, range = c(-0.1, 0.1), precision = 0.5)
+  testthat::expect_equal(xprope$mhdior, c(0.695, 1, 0.87), tolerance = 0.1)
 })
 
 test_that("emmGrid point_estimate", {
@@ -78,7 +80,7 @@ test_that("emmGrid bayesfactor_parameters", {
   set.seed(4)
   xsdbf <- bayesfactor_parameters(all_, prior = model)
   testthat::expect_equal(log(xsdbf$BF), c(-2.5756125848835, 1.69713280431204, -0.212277519930343), tolerance = .1)
-  testthat::expect_warning(bayesfactor_savagedickey(all_))
+  testthat::expect_warning(bayesfactor_parameters(all_))
 })
 
 test_that("emmGrid bayesfactor_restricted", {
@@ -109,4 +111,43 @@ test_that("emmGrid describe_posterior", {
   )
   testthat::expect_equal(log(xpost$BF), c(-2.58, 2.00, -0.25), tolerance = 0.1)
   testthat::expect_warning(describe_posterior(all_, test = "bf"))
+})
+
+## For non linear models
+set.seed(333)
+df <- data.frame(
+  G = rep(letters[1:3], each = 2),
+  Y = rexp(6)
+)
+
+fit_bayes <- stan_glm(Y ~ G,
+  data = df,
+  family = Gamma(link = "identity"),
+  refresh = 0
+)
+fit_bayes_prior <- update(fit_bayes, prior_PD = TRUE)
+
+bayes_sum <- emmeans(fit_bayes, ~G)
+bayes_sum_prior <- emmeans(fit_bayes_prior, ~G)
+
+test_that("emmGrid bayesfactor_restricted2", {
+  testthat::skip_on_travis()
+  testthat::skip_on_cran()
+
+  hyps <- c("a < b", "b < c")
+  xrbf1 <- bayesfactor_restricted(bayes_sum, fit_bayes, hypothesis = hyps)
+  xrbf2 <- bayesfactor_restricted(bayes_sum, bayes_sum_prior, hypothesis = hyps)
+
+  testthat::expect_equal(xrbf1, xrbf2, tolerance = 0.1)
+})
+
+
+test_that("emmGrid bayesfactor_parameters", {
+  testthat::skip_on_travis()
+  testthat::skip_on_cran()
+
+  xsdbf1 <- bayesfactor_parameters(bayes_sum, prior = fit_bayes)
+  xsdbf2 <- bayesfactor_parameters(bayes_sum, prior = bayes_sum_prior)
+
+  testthat::expect_equal(log(xsdbf1$BF), log(xsdbf2$BF), tolerance = 0.1)
 })