+++ NEWS.md
@@ -1,14 +1,21 @@
-# bayestestR 0.4.1
+# bayestestR 0.4.9
 
 ## General
 
 - Added `plot()`-method for `p_significance()`.
 - `p_significance()` now also works for *brmsfit*-objects.
 - `equivalence_test()` gets `effects` and `component` arguments for *stanreg* and *brmsfit*  models, to print specific model components.
+- Support for *mcmc* objects (package **coda**=)
+- Provide more distributions via `distribution()`.
+
+## Breaking changes
+
+- `p_rope()`, previously as exploratory index, was renamed as `mhdior()` (for *Max HDI inside/outside ROPE*), as `p_rope()` will refer to `rope(..., ci = 1)` ( #258 )
 
 ## Bug fixes
 
 - Fixed mistake in description of `p_significance()`.
+- Fixed error when computing BFs with `emmGrid` based on some non-linear models ( #260 ).
 
 # bayestestR 0.4.0